fix: harden cache rows and embedded parses, type coordinator unavailability
Round-2 quality pass, eight findings from a structural survey:

Corrupt or legacy D1 cache rows (invalid JSON, wrong-shaped body or
headers) now decode as cache misses on both the fresh and stale read paths
instead of throwing a relay 500 -- one bad row could previously turn a
cached route into an error for every caller.

Embedded GitHub page connections (labels, assignees, actors) previously
accepted absent or malformed pageInfo and could synthesize a partial list;
they now require pageInfo.hasNextPage === false and otherwise fall through
to the exact API.

The pool coordinator returns a typed unavailable-selection result instead
of throwing across the Durable Object RPC boundary; the relay translates
it to the same identities_cooling_down flow, the fragile error-message
string match is gone, and the "uncaught exception" noise in every e2e run
disappears. The never-read SelectionRequest.pool field and the unreachable
"fallback" selection reason are deleted, with docs corrected.

Also: requestTimeoutMs() replaces ten duplicated timeout parses;
generate-routes.mjs validates every output before writing any tracked
file; migration 0015 drops the orphaned oauth_states table (OAuth state
moved to signed cookies long ago); the 999-line actions-cache e2e file is
split along its three existing describe boundaries.
